requirements.txt
PyYAML==6.0.1
scikit-learn==1.5.0
gensim==4.3.2
scipy==1.10.1